STIR: Renamed circle_stir to stir and circle_fft to fft to allow better diff resolution
